Refactor ActivityIngestionMetaCheck companions
